Stop using 'NodeGit.Submodule.lookup'.
It's extremely slow and doesn't seem to parallelize.  There are still two
places where it's being used: in 'open' as we need the 'init' method, and in
'status' to get the URL of newly added submodules; the latter use can be fixed
after adjusting code to parse the .gitmodules file to also get urls.