Reworked 'push' to work when pushing to new branch
This fix required quite a big change.  Previously I was determining which
submodules needed to be pushed by looking at the difference between the source
and target branches in the meta-repo: when the target branch was new this
implied all commits, and implied that all submodules much be pushed.  We would
therefore fail when any submodules were closed.

Now I am determining which commits on the source branch do not exist in the
remote, and am computing the submodules that must be pushed based on these
commits, regardless of target branch.