Cleanup unreachable commits
Due to the way I use a single commit repo to provide one commit namespace, each
repo ends up with all commits in the multi-repo namespace.  This allowed a bug
to slip through in `merge` where the `metaRepo` was referenced instead of a
submodule's repo, and the command worked (in test) anyway because the metaRepo
had all the commits from the submodule.