Reworked git-meta documentation.
This is the first step in changing the mission of git-meta to be a mechanism
for implementing a mono-repo.

I have trimmed down the `README.md` file to provide a basic introduction and
quick-start for git-meta.  I've split out an architectural whitepaper into
`doc/architecture.md` and added references in `README.md` to future documents
such as user and administration guides.