expectation transform in testMultiRepoManipulator
Changed `testMultiRepoManipulator` to allow the caller to transformn the map of
expected ASTs immediately prior to checking them against the actual result.
This capability is necessary to test cases where the actual state of the
repository may contain generated (mapped) values where the system cannot adjust
them -- for example, if a branch (or ref) name has the ID of a commit embedded
in it.