Separate concerns of `bare` and `head`.
I made a poor decision early on to let a null head imply a bare repo in the
test framework.  Now, I need to change some parts of the code to handle
headless repositories, and I'd like to be able to reflect such a state in the
test framework.  This change decouples the two concepts of bareness and
headlessness.