support for prompting for commit message
When "-m" is not specified, `git-meta commit` will now open the configured
editor and allow the user to edit the message.  I had to extend and tweak the
status logic because previously, we were relying on the result of
`Commit.commit` to determine (and warn) if no changes were available.  Now we
have to check ourselves before we (potentially) bring up the editor.