Fix when onto commits have submodule change.
If the commits being rebased onto have changes to an open submodule, at the end
of a rebase operation we were leaving the HEAD of such submodules on the wrong
commit.  This situation would have left those submodules in a bad state, but it
also caused the rebase commit generation to fail.

Fixes: https://github.com/twosigma/git-meta/issues/172