Fixed getSubmoduleStatus to check workdir vs. tree
Changed `getSubmoduleStatus` to ignore the commit staged in the index for a
submodule if that submodule has an open repository.  When a submodule has an
open repository, we -- in keeping with the general model of git-mta -- reflect
the status of the workdir HEAD as (automatically) staged, as measured against
the commit registered in the HEAD of the meta-repo.  For example, this change
fixes the problem where we could end up making an empty meta-repo commit (and
give misleading status information) when a submodule has a new commit staged in
the index and an old commit on HEAD in its open repo.

I also factored out a `getRelation` method from `getSubmoduleStatus` to make
that functionality easier to test and reduce the size of `getSubmoduleStatus`.

Finally, I changed `amendRepo` and `commit` to preemptively stage all
submodules before making a meta-repo commit; otherwise, we might inadvertantly
commit a staged submodule commit change without also getting a reversion from
the workdir.