Support workdir change in listSubmoduleDescriptors
Generally, submodule commit changes are auto-staged.  However, in some
situations -- like path-based commits and amends -- we do not auto-commit all
submodule modifications.  Therefore, if we do have unstaged submodule commits
registered in a `RepoStatus` method, we need this method to handle them so that
we can print an appropriate message as a prompt in the editor.