Improve `amend` to allow it to work in more states
Also, status display in the editor prompt is much improved.

Previously `amend` would refuse to run in many situations (e.g., new commits in
a repo previously modified in the meta-repo commit being amended).  Now, it
will refuse only when the signatures of the sub-module commits referenced in
the last repo do not match; this is because we would somewhat blindly allow
their commit messages to be overwritten.

This change also paves the way for `--interactive` mode to work with amend,
which will be the recommended way to apply an amend in the one situation where
it doesn't work now -- the interactive version will allow you to edit different
commit messages separately.