Speedup `getRepoStatus` and clients.
The status functions are only supposed to report changed/interesting
things, but it was loading a `RepoAST.Submodule` object for each
submodule, even those not open and unchanged.  Now, it will process open
submodules and those with changes between the index and last commit.
The latter information will be determined via `DiffUtil`.

This change propagated through several clients of `getRepoStatus` who
assumed that the `submodules` property of the returned `RepoStatus`
object contained *all* current submodules.  Now it contains only the
interesting (changed or open) submodules.