Fixed problem with cleaning up conflicts.
I'm not sure why this bug was not replicated in my tests, but it appears
that calling `Index.conflictCleanup` *after* writing the index can cause
it to be corrupted, preventing the commit from being written and leaving
the repository in a bad state (`git status` complains that there are
multiple stage entries).

To fix, I've put conflict removals for each path that has been
addressed, rather than relying on the bulk cleanup (except in the case
of `abort`).