Fixed bug where reverted files were not committed
https://github.com/twosigma/git-meta/issues/473

When doing an amend, we generate a special 'RepoStatus' object that
describes what the resulting commit will actually look like.  This
allows us to display the right thing in the messsage prompt, and to know
when no change is being made (i.e., it's a reversion).

However, we were also staging exactly the files flagged as "staged" in
the status object.  Since reverted files are missing from this object,
they were not being staged, and thus were ignored when they shouldn't
have been -- in 'all' mode.  I changed the logic to just stage all files
when in 'all' mode.