Fix flaky `StashUtil.removeStash`
It was missing a yield statement to where I was pointing the stash
reference to a new commit.  When the yield is added, the test failed
because we ended up with a duplicate entry in the reflog.  I believe
that what was happening when the test succeeded was that the reference
itself was updated but that the program exited before the reflog entry
was added.

After adding the `yield`, I had to add another statement to remove the
duplicate entry.

I validated that the test no longer flaked by running it in a loop in
bash.  Previously it would fail every second or third time.