Stop `reset --hard` if will remove dirty open subs
In this case, libgit2 will delete the submodule's directory, throwing
away all changes.  See https://github.com/twosigma/git-meta/issues/509

I do not want to do this check when resetting to abort cherry-pick -- we
started out clean before picking and it could be pretty common that
we've added a submodule that we want to get rid of when we abort -- but
I don't want to add a weird flag that will exist just for this
workaround.  So, I went ahead and added a `MERGE` reset type, that will
eventually do a real `git reset --merge` but for the time being is just
like `HARD` except that it does not do this check.