Ditch `upstream` if would rewrite too many commits
When you specify an upstream to libgit2, it attempts to rewrite all
commits between your current commit and `upstream`, even those in
`onto`.  It fails with these commits in an unusual way, not by giving an
error in the rebsase operation, but when you try to make a rebase commit
(probably because the commit would be empty) it doesn't actually
succeed, but also doesn't generte an error; instead, it fails and
returns a garbage commit sha (probably the failure to throw an error is
a flaw in nodegit).

I only detected this situation after tightening up the test harness, so that
these garbage commit shas are flagged.  They may be harmless, but it
seems wrong (and I could see the attempt to rewrite these commits
creatng conflicts or non-empty commits), so I'm changing the logic to
ignore upstream when it's equal to or an ancestor of HEAD.