sync sparse-checkout & index after open/close
The current strategy for sparse mode was to exclude in the
sparse-checkout file and index all files besides `.gitmodules`.  This
strategy was fine as far as `git-meta` was concerned generally, but it
confused Git; since `git-meta` defers to Git for many commands (e.g.,
`diff`), and since we generally want to allow people to use normal Git
commands and maintain our repo in a recognizable state, the current
strategy is not sufficient.

Instead, we need to maintain the invarian that "open" submodules are
seen as visible in the sparse mode (not skipped) and "closed" submodules
are seen as invisible (skipped).

Severl changes were needed to make this happen:

* We now consider a repo to be "inSparseMode" if the config is set and
  don't also check to see if `".gitmodules\n"` is the contents of the
  `sparse-checkout` file.  It was a little pedantic to add that to the
  check to begin with, and now we're going to have an arbitrary set of
  paths.
* When opening a submodule, we add its path to `sparse-checkout`.
* `SubmoduleConfigUtil.deinit` now removes submodules from
  `sparse-checkout`.  Since this would be O(N^2) (e.g. if you're closing
  all submodules) and in all cases we generally know up front the list
  of subs to remove, `deinit` now takes a list of subs to deinitialize.
* Opening and closing subs now requires an update to the index, since
  libgit2 (and I'm guessing not even Git) will do what we need by
  default.  To facilitate this, I updated the contracts of the
  open/close methods, and moved `writeMetaIndex` down into
  `SparseCheckoutUtil`.
* Call `writeMetaIndex` from stash_util, which didn't need to update the
  index previously
* And also from `CloseUtil`.