Use libgit2 to compute submodule tree changes
Addresses: https://github.com/twosigma/git-meta/issues/595

Previously, I was manually computing what submodule changes and
conflicts would result from a merge or cherry-pick (rebase uses
cherry-pick).  In the case of a merge, I used "the" merge base between
the head and target commits -- however, there may be many merge bases,
and the logic used to determine what to do is beyond the scope of
git-meta.  I started doing this instead of the obvious `Merge.merge` or
`CherryPick.cherryPick` command due to the littany of problems I have
using libgit2 in the workdir of the meta-repo (performance, correctness
e.g., with sparse-checkouts, etc.).

With this change, I go back to using libgit2 to do this computation, but
I'm using the "commits" versions of these commands -- `Merge.commits`
and `CherryPick.commits` -- which do not touch the worktree, but work
only on commits.  Now, the only manual deduction I do is to determine
which submodule conflicts flagged by libgit2 are candidates for
resolution (via merge/cherry-pick in the affected submodule).

As can be seen from the test case added to `merge`, it does solve the
problem.