Better logic to compute which submodules to push
The previous logic was to rule out pushing submodules if their current
commits (or their descendants) existed in one of a set of relevant
branches (e.g., tracking branches).  The downside of this strategy was
that we couldn't deduce parent-child relationships of submodule commits
that didn't exist locally.

The new strategy is to include as candidates only submodules that were
not changed between the commit being pushed and any of the (merge bases
of the) candidate branches.

Also fixed a bug where we were not constructing the ref name of a
tracking branch correctly.