Pull: do not check remote name
We don't care if the remote they give is a *named* remote -- we just
care if we can fetch from it.  Fetching calls out to regular git,
which knows how to handle not only remote names but also remote URLs.
And fetching is about the first thing we do, so we'll find out quickly
whether the remote is valid.