Fix up-to-date case in git merge-bare.
In a normal merge (merge "theirs" commit into "ours"), if "ours" is a descendant of "theirs" then it is considered as up-to-date case. If "theirs" is a descedant of "ours", then it is a fast-forward merging.

In merge-bare mode, "theirs" and "ours" are symetrical, and git-meta should not exit early as if it is an up-to-date case