stitch: prefix ref names for submodule refs
When fetching from a submodule, we would fetch into a global synthetic
ref namespace.  This works, but it makes ref negotiation take
a long time.

Instead, we figure that there is little to no overlap between submodules,
and prefix submodule refs with the name of their submodule.

Note that this could, in theory, lead to d/f conflicts if someone names
a submodule after a SHA, but this would require true malice.