stitch: Handle missing converted commit
Handle the case where a notes is present but the converted commit to is not
present.  This can happen if you are stitching and pushing notes in parallel:
it's possible that you push a note for a commit that doesn't end up getting
pushed.  That would cause later conversions which fetch the notes to fail,
unless we handle that case.